refactor: move chat delivery inlet to agent runtime
