+++ backend/web/services/chat_delivery_hook.py
@@ -1,5 +1,5 @@
-"""Compatibility shell for messaging runtime delivery bridge."""
+"""Compatibility shell for Agent Runtime chat inlet."""
 
-from messaging.delivery.runtime_bridge import make_chat_delivery_fn
+from backend.agent_runtime.chat_inlet import make_chat_delivery_fn
 
 __all__ = ["make_chat_delivery_fn"]
