+++ messaging/delivery/contracts.py
@@ -1,0 +1,23 @@
+"""Pure contracts for Chat delivery wiring."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Any, Protocol
+
+
+@dataclass(frozen=True)
+class ChatDeliveryRequest:
+    recipient_id: str
+    recipient_user: Any
+    content: str
+    sender_name: str
+    sender_type: str
+    chat_id: str
+    sender_id: str
+    sender_avatar_url: str | None
+    signal: str | None
+
+
+class ChatDeliveryFn(Protocol):
+    def __call__(self, request: ChatDeliveryRequest) -> None: ...
